# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "onec",
    "version": "2.0.0"
  },
...
  "paths": {
...
    "/negotiation/{id}": {
      "put": {
        "operationId": "updateNegotiation",
        "tags": [
          "negotiation"
        ],
        "description": "Update a Negotiation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "nullable": true
                  },
                  "client": {
                    "type": "string",
                    "nullable": true
                  },
                  "user": {
                    "type": "string",
                    "nullable": true
                  },
                  "tags": {
                    "type": "string",
                    "nullable": true
                  },
                  "step": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "nullable": true
                  },
                  "value": {
                    "type": "number",
                    "nullable": true
                  },
                  "startsDate": {
                    "type": "string",
                    "nullable": true
                  },
                  "partnerId": {
                    "type": "string",
                    "nullable": true
                  },
                  "observation": {
                    "type": "string",
                    "nullable": true
                  },
                  "averageGuide": {
                    "type": "number",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "negotiation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string",
                          "nullable": true
                        },
                        "client": {
                          "type": "string",
                          "nullable": true
                        },
                        "user": {
                          "type": "string",
                          "nullable": true
                        },
                        "tags": {
                          "type": "string",
                          "nullable": true
                        },
                        "step": {
                          "type": "string",
                          "nullable": true
                        },
                        "status": {
                          "type": "string",
                          "nullable": true
                        },
                        "value": {
                          "type": "number",
                          "nullable": true
                        },
                        "startsDate": {
                          "type": "string",
                          "nullable": true
                        },
                        "observation": {
                          "type": "string",
                          "nullable": true
                        },
                        "partnerId": {
                          "type": "string",
                          "nullable": true
                        },
                        "averageGuide": {
                          "type": "number",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "nullable": true
                        },
                        "updatedAt": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "client",
                        "user",
                        "tags",
                        "step",
                        "status",
                        "value",
                        "startsDate",
                        "observation",
                        "partnerId",
                        "averageGuide",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "negotiation"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteNegotiation",
        "tags": [
          "negotiation"
        ],
        "description": "Delete a negotiation",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getNegotiationById",
        "tags": [
          "negotiation"
        ],
        "description": "Get a list of Negotiation by id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string",
                        "nullable": true
                      },
                      "client": {
                        "type": "string",
                        "nullable": true
                      },
                      "user": {
                        "type": "string",
                        "nullable": true
                      },
                      "tags": {
                        "type": "string",
                        "nullable": true
                      },
                      "step": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string"
                      },
                      "value": {
                        "type": "number",
                        "nullable": true
                      },
                      "startsDate": {
                        "type": "string",
                        "nullable": true
                      },
                      "observation": {
                        "type": "string",
                        "nullable": true
                      },
                      "averageGuide": {
                        "type": "number",
                        "nullable": true
                      },
                      "partnerId": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "title",
                      "client",
                      "user",
                      "tags",
                      "step",
                      "status",
                      "value",
                      "startsDate",
                      "observation",
                      "averageGuide",
                      "partnerId"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}